Move include files into their own folder
Moves the include files into a seperate file to
source files.

Fix lowercase issues with inc folder
