+++ CorsixTH/inc/cp936_table.h
