+++ CorsixTH/inc/run_length_encoder.h
