+++ CorsixTH/inc/th_sound.h
